### 2019-03-22 #1343 @whedon
[PRE REVIEW]: Container Tree: Software to Model Container Filesystems, Packages, and Inheritance
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** <a href="https://www.github.com/singularityhub/container-tree" target ="_blank">https://www.github.com/singularityhub/container-tree</a>
**Version:** 0.0.48
