### 2019-03-22 #1343 @arfon
> A new robot! ❤️

Yes! This one [isn't ours though](https://github.com/swinton/probot-ooo). @vsoch - would you be willing to wait until `@trallard` is back before moving this forward?
