### 2019-04-27 #1343 @vsoch
Hey guys are we ready to move forward?
