### 2019-04-30 50bada0
adding reference to docs in first paragraph
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
