+++ paper/paper.bib
@@ -11,6 +11,13 @@
   address   = "Houston, TX"
 }
 
+@MISC{container-tree,
+  title       = "container-tree",
+  abstract    = "container-tree: data structures to study containers",
+  howpublished = "\url{https://github.com/singularityhub/container-tree}",
+  institution = "Github"
+}
+
 @MISC{container-diff,
   title       = "container-diff",
   abstract    = "container-diff: Diff your Docker containers",
