### 2019-04-30 15518f3
adding dates for "nd" entries
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
