+++ paper/paper.bib
@@ -15,6 +15,7 @@
   title       = "container-tree",
   abstract    = "container-tree: data structures to study containers",
   howpublished = "\url{https://github.com/singularityhub/container-tree}",
+  year      =  2018,
   institution = "Github"
 }
 
@@ -22,6 +23,7 @@
   title       = "container-diff",
   abstract    = "container-diff: Diff your Docker containers",
   howpublished = "\url{https://www.github.com/GoogleContainerTools/container-diff}",
+  year      =  2017,
   institution = "Github"
 }
 
