### 2019-04-30 #1418 @vsoch
okay, looks good! Thanks @whedon. I had some "n.d" dates that I wanted to fix up. The manuscript should be ready now. Thanks for reviewing @luizirber and overseeing @brainstorm.
