### 2019-05-01 4950a26
fix local install instructions
