### 2019-05-05 #1418 @luizirber
Not ready yet, pycon happened along the way... will finish tomorrow
