### 2019-05-05 #1418 @vsoch
Yeah @luizirber that sounds awesome!! When you come back, let us know how it was, and where PyCon 2020 will be! I found [the site](https://us.pycon.org/2019/about/) and I didn't realize it was closer to my neck of the woods than say, San Francisco. I totally want to go one of these years, so fingers crossed it's close by again!
