### 2019-05-05 #1418 @vsoch
omg it's in Pittsburgh! https://pycon.blogspot.com/2019/02/pycon-2020-2021-location.html woo!
