+++ docs/_docs/contributing/index.md
@@ -7,8 +7,8 @@ order: 1
 
 It's so great that you want to contribute! There are several ways to contribute.
 
- - [Contribute to Documentation]({{ site.url }}{{ site.baseurl }}/contributing/docs/)
- - [Contribute Code]({{ site.url }}{{ site.baseurl }}/contributing/code/)
+ - [Contribute to Documentation]({{ site.baseurl }}/contributing/docs/)
+ - [Contribute Code]({{ site.baseurl }}/contributing/code/)
  - Ask a question, report a bug, or suggest a feature by opening up an [issue](https://www.github.com/{{ site.repo }}/issues)
 
 Whether you ask a question, help to update the code with a bug fix, or
