### 2019-05-13 #1418 @luizirber
Finally back on track, sorry for the delay.

I sent some PRs for small doc fixes in the original repo, so the main items missing in the checklist are:

## Automated tests

While CircleCI is set up, I didn't find any tests in the codebase. The examples are very nice, but are not automated/ran on every PR. But it seems they would be very amenable to testing, perhaps with smaller data inputs?

## References

At least the PLOS ref has a DOI. Not sure about the others.
