### 2019-05-13 #1418 @vsoch
Thanks @luizirber ! I'll get you some tests (likely tomorrow, or day after). I just responded to the PR (thank you!) and I'll see if I can update references with DOIs.
