+++ paper/paper.bib
@@ -8,7 +8,8 @@
   number    =  239,
   month     =  mar,
   year      =  2014,
-  address   = "Houston, TX"
+  address   = "Houston, TX",
+  url       = {https://www.linuxjournal.com/content/docker-lightweight-linux-containers-consistent-development-and-deployment}
 }
 
 @MISC{container-tree,
@@ -16,7 +17,8 @@
   abstract    = "container-tree: data structures to study containers",
   howpublished = "\url{https://github.com/singularityhub/container-tree}",
   year      =  2018,
-  institution = "Github"
+  institution = "Github",
+  doi       = {10.5281/zenodo.2602127}
 }
 
 @MISC{container-diff,
@@ -60,5 +62,6 @@
   pages    = "e0188511",
   month    =  nov,
   year     =  2017,
-  language = "en"
+  language = "en",
+  doi      = "{https://doi.org/10.1371/journal.pone.0188511}"
 }
