### 2019-05-14 54bc0eb
doi should not be in quotes
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
