+++ paper/paper.bib
@@ -8,8 +8,7 @@
   number    =  239,
   month     =  mar,
   year      =  2014,
-  address   = "Houston, TX",
-  url       = {https://www.linuxjournal.com/content/docker-lightweight-linux-containers-consistent-development-and-deployment}
+  address   = "Houston, TX"
 }
 
 @MISC{container-tree,
@@ -63,5 +62,5 @@
   month    =  nov,
   year     =  2017,
   language = "en",
-  doi      = "{https://doi.org/10.1371/journal.pone.0188511}"
+  doi      = {https://doi.org/10.1371/journal.pone.0188511}
 }
