### 2019-05-14 551ada1
adding start to testing, needs a lot more work, just saving in case my computer explodes
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
