### 2019-05-14 92af8c8
adding tests for collection (still more to go!)
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
