### 2019-05-14 d96fb60
typo in config
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
