### 2019-05-15 #1418 @brainstorm
@vsoch LGTM, archive to Zenodo #FTW?
