### 2019-05-15 ffc7dd9
adding all the tests!
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
