### 2019-05-15 9f19685
rename container diff to not have extension
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
