### 2019-05-15 b7f545a
try moving to /usr/local/bin
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
