### 2019-05-15 55db1d7
adding container-diff to path
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
