### 2019-05-15 01d0ec5
try including Dockerfiles
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
