+++ MANIFEST.in
@@ -1,6 +1,7 @@
 include README.md LICENSE
 recursive-include containertree *
 recursive-include containertree *.html
+recursive-include containertree Dockerfile
 global-exclude *.py[co]
 global-exclude *_extra.py
 global-exclude *.img
