### 2019-05-15 4f97e57
need to include Dockerfile with tests
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
