### 2019-05-15 5d066f8
forgot to add Dockerfile"
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
