+++ containertree/server.py
@@ -15,11 +15,15 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 
-import http.server
 import socketserver
 from random import choice
 from threading import Thread
 import os
+
+try:
+    import http.server as http_server
+except:
+    from BaseHTTPServer import BaseHTTPRequestHandler as http_server
 
 
 def get_web_server(port=None):
@@ -33,7 +37,7 @@ def get_web_server(port=None):
     if port == None:
         port = choice(range(2000,9999))
 
-    Handler = http.server.SimpleHTTPRequestHandler
+    Handler = http_server.SimpleHTTPRequestHandler
     httpd = socketserver.TCPServer(("", port), Handler)
     server = Thread(target=httpd.serve_forever)
     server.setDaemon(True)
