+++ containertree/server.py
@@ -15,16 +15,16 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 
-import socketserver
 from random import choice
 from threading import Thread
 import os
 
 try:
     import http.server as http_server
+    import socketserver
 except:
     from BaseHTTPServer import BaseHTTPRequestHandler as http_server
-
+    import SocketServer as socketserver
 
 def get_web_server(port=None):
     '''get_web_server returns a httpd object (socket server)
