### 2019-05-15 d42fd7e
ensure python3 being used
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
