### 2019-05-15 0b92dca
separate tests for python2/3
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
