+++ .circleci/config.yml
@@ -111,8 +111,20 @@ install_python_2: &install_python_2
            echo "Miniconda 2 is already installed, continuing to build."
        fi
 
-test_containertree: &test_containertree
-  name: Test Containertree
+test_containertree3: &test_containertree3
+  name: Test Containertree Python3
+  command: |
+        export PATH=$HOME/conda/bin:$PATH
+        python3 --version
+        cd ~/repo/containertree/tests
+        $HOME/conda/bin/python3 -m unittest test_utils
+        $HOME/conda/bin/python3 -m unittest test_collection
+        $HOME/conda/bin/python3 -m unittest test_container
+        $HOME/conda/bin/python3 -m unittest test_packages
+        /bin/bash test_client.sh
+
+test_containertree2: &test_containertree2
+  name: Test Containertree Python2
   command: |
         export PATH=$HOME/conda/bin:$PATH
         python --version
@@ -122,6 +134,7 @@ test_containertree: &test_containertree
         $HOME/conda/bin/python -m unittest test_container
         $HOME/conda/bin/python -m unittest test_packages
         /bin/bash test_client.sh
+
 
 jobs:
   test-python-3:
@@ -139,7 +152,7 @@ jobs:
           paths:
             - /home/circleci/conda
           key: v1-dependencies
-      - run: *test_containertree
+      - run: *test_containertree3
 
   test-python-2:
     machine: true
@@ -156,4 +169,4 @@ jobs:
           paths:
             - /home/circleci/conda
           key: v1-dependencies        
-      - run: *test_containertree
+      - run: *test_containertree2
