### 2019-05-15 6ba4f1e
updating links to python 3
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
