+++ .circleci/config.yml
@@ -87,11 +87,10 @@ install_python_3: &install_python_3
   command: | 
       ls $HOME
       if [ ! -d "/home/circleci/conda" ]; then
-          wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
+          wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh
           /bin/bash Miniconda3-latest-Linux-x86_64.sh -b -p $HOME/conda
           export PATH=$HOME/conda/bin:$PATH
           $HOME/conda/bin/conda install pandas
-          cp $HOME/conda/bin/python3 $HOME/conda/bin/python
           $HOME/conda/bin/python setup.py install
        else
            echo "Miniconda 3 is already installed, continuing to build."
@@ -111,20 +110,8 @@ install_python_2: &install_python_2
            echo "Miniconda 2 is already installed, continuing to build."
        fi
 
-test_containertree3: &test_containertree3
-  name: Test Containertree Python3
-  command: |
-        export PATH=$HOME/conda/bin:$PATH
-        python3 --version
-        cd ~/repo/containertree/tests
-        $HOME/conda/bin/python3 -m unittest test_utils
-        $HOME/conda/bin/python3 -m unittest test_collection
-        $HOME/conda/bin/python3 -m unittest test_container
-        $HOME/conda/bin/python3 -m unittest test_packages
-        /bin/bash test_client.sh
-
-test_containertree2: &test_containertree2
-  name: Test Containertree Python2
+test_containertree: &test_containertree
+  name: Test Containertree
   command: |
         export PATH=$HOME/conda/bin:$PATH
         python --version
@@ -134,7 +121,6 @@ test_containertree2: &test_containertree2
         $HOME/conda/bin/python -m unittest test_container
         $HOME/conda/bin/python -m unittest test_packages
         /bin/bash test_client.sh
-
 
 jobs:
   test-python-3:
@@ -152,7 +138,7 @@ jobs:
           paths:
             - /home/circleci/conda
           key: v1-dependencies
-      - run: *test_containertree3
+      - run: *test_containertree
 
   test-python-2:
     machine: true
@@ -169,4 +155,4 @@ jobs:
           paths:
             - /home/circleci/conda
           key: v1-dependencies        
-      - run: *test_containertree2
+      - run: *test_containertree
