+++ .circleci/config.yml
@@ -130,14 +130,14 @@ jobs:
       - checkout
       - restore_cache:
           keys:
-          - v1-dependencies
+          - v2-dependencies
       - run: *install_python_3
       - run: *install_containerdiff
       - run: *install
       - save_cache:
           paths:
             - /home/circleci/conda
-          key: v1-dependencies
+          key: v2-dependencies
       - run: *test_containertree
 
   test-python-2:
@@ -147,12 +147,12 @@ jobs:
       - checkout
       - restore_cache:
           keys:
-          - v1-dependencies
+          - v2-dependencies
       - run: *install_python_2
       - run: *install_containerdiff
       - run: *install
       - save_cache:
           paths:
             - /home/circleci/conda
-          key: v1-dependencies        
+          key: v2-dependencies        
       - run: *test_containertree
