### 2019-05-15 d220f6e
Add tests (#44)
* adding start to testing, needs a lot more work, just saving in case my computer explodes

Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>

* adding tests for collection (still more to go!)

Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>

* typo in config

Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>

* adding all the tests!

Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>

* rename container diff to not have extension

Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>

* try moving to /usr/local/bin

Signed-off-by: Vanessa Sochat <vsochat@stanford.e
[…]
