+++ containertree/utils/__init__.py
@@ -25,6 +25,9 @@ from .fileio import (
     read_file,
     recursive_find,
     run_command,
+    print_json,
+    write_file,
+    write_json,
     check_install    
 )
 
