+++ MANIFEST.in
@@ -1,4 +1,4 @@
-include README.md LICENSE
+include README.md LICENSE Dockerfile
 recursive-include containertree *
 recursive-include containertree *.html
 global-exclude *.py[co]

[…]
