### 2019-05-15 #1418 @vsoch
@luizirber I'm so glad you noticed this, because I had added the circle config to build containers without finishing the actual tests for the trees and client! I'm happy to report that I now have tests for container trees, collection trees, package trees, data export, and the command line client! All just under 1K lines of code, tested and merged: https://github.com/singularityhub/container-tree/pull/44

@brainstorm I've drafted a new release for the updated version, v0.0.49, and the archive should be processing --> available via zenodo asap https://zenodo.org/account/settings/github/repository/singularityhub/container-tree#

Holy crap, I'm tired... going for a run! :running_man:
