### 2019-05-18 #1418 @brainstorm
Cool cool, thanks everybody! @vsoch need the zenodo DOI thingie and should be Archivable, I cannot login/hack your zenodo account ;)
