### 2019-05-18 #1418 @vsoch
And latest release https://github.com/singularityhub/container-tree/releases/tag/v0.0.49
