### 2019-05-19 #1418 @labarba
@vsoch — Please edit the metadata of the Zenodo deposit (no need to get new version or new DOI) so the title and author list match the JOSS paper. You may also want to add your ORCID in the metadata.
