### 2019-05-19 #1418 @vsoch
It looks like the pdf is 404, do we need to wait longer? https://www.theoj.org/joss-papers/joss.01418/10.21105.joss.01418.pdf
