### 2019-05-19 #1418 @vsoch
Showing up for me too! Thanks @arfon. And huge +1, GitHub pages is slow these days. Thanks to everyone for your help with this paper! :fireworks:
